requirements.txt
# AI Papers Tracker - Python 依赖

# HTTP 客户端
httpx>=0.24.0
requests>=2.31.0

# HTML 解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据库
pymysql>=1.1.0
cryptography>=41.0.0

# 定时任务
APScheduler>=3.10.0

# 配置管理
python-dotenv>=1.0.0

# 工具库
tqdm>=4.65.0

# X Trending 数据源 (xAI Grok API)
xai-sdk>=1.5.0
pydantic>=2.5.0
